Add slurm-alternative task scheduler
./swarm/evaluate-test-worker.py is a drop-in alternative to
swarm/evaluate-test.sh which launches worker tasks which can take on
multiple evaluation tasks, to get around slow backfilling. Also can
resume from an existing report, and can do so proof-wise when passed the
--mode="proof" flag. Also accepts a --num-workers=<n> flag which sets
the number of worker tasks, defaults to 32. Probably shouldn't set it
too high because the work queue is synchronized using the filesystem, so
not super thread safe. 128 appears to be generally safe.